[vcpkg] Remove vcpkg::Archives from vcpkg::base. Add vcpkg::Hash.
