+++ toolsrc/include/vcpkg/commands.h
@@ -131,9 +131,6 @@ namespace vcpkg::Commands
 
     namespace Hash
     {
-        std::string get_string_hash(const std::string& s, const std::string& hash_type);
-        std::string get_file_hash(const Files::Filesystem& fs, const fs::path& path, const std::string& hash_type);
-
         void perform_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
     }
