+++ toolsrc/src/vcpkg/commands.cpp
@@ -1,5 +1,7 @@
 #include "pch.h"
 
+#include <vcpkg/base/hash.h>
+
 #include <vcpkg/build.h>
 #include <vcpkg/commands.h>
 #include <vcpkg/export.h>
@@ -75,6 +77,29 @@ namespace vcpkg::Commands::Fetch
         const std::string tool = args.command_arguments[0];
         const fs::path tool_path = paths.get_tool_exe(tool);
         System::println(tool_path.u8string());
+        Checks::exit_success(VCPKG_LINE_INFO);
+    }
+}
+
+namespace vcpkg::Commands::Hash
+{
+    const CommandStructure COMMAND_STRUCTURE = {
+        Strings::format("The argument should be a file path\n%s",
+                        Help::create_example_string("hash boost_1_62_0.tar.bz2")),
+        1,
+        2,
+        {},
+        nullptr,
+    };
+
+    void perform_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths)
+    {
+        Util::unused(args.parse_arguments(COMMAND_STRUCTURE));
+
+        const fs::path file_to_hash = args.command_arguments[0];
+        const std::string algorithm = args.command_arguments.size() == 2 ? args.command_arguments[1] : "SHA512";
+        const std::string hash = vcpkg::Hash::get_file_hash(paths.get_filesystem(), file_to_hash, algorithm);
+        System::println(hash);
         Checks::exit_success(VCPKG_LINE_INFO);
     }
 }
